实现抢购神券功能

+++ src/main/java/com/example/entity/dto/resp/TicketRespDTO.java
@@ -24,9 +24,9 @@ public class TicketRespDTO {
     Integer validDateType;
     Date validDate;
     @Min(0)
-    Double price;
+    Float price;
     @Min(0)
-    Integer count;
+    Long count;
 
     Date createTime;
